dom/base/test/jsmodules/test_scriptInsertedModule.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Test a script-inserted module</title>
<script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
<script>
  var result;

  let script = document.createElement("script");
  script.type = "module";
  script.src = "./module_simpleImport.mjs";
  document.head.appendChild(script);

  SimpleTest.waitForExplicitFinish();

  function testLoaded() {
    ok(result == 42);
    SimpleTest.finish();
  }
</script>
<body onload='testLoaded()'></body>
